athena-backend/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
Pillow
google-generativeai
python-dotenv
